tools/bridge-monitor/host/requirements.txt
paho-mqtt>=1.6
pyserial>=3.5

